Rename test files and split delegate-retention test by store
Drop the RCAdMob* prefix from all adapter test files and class names so
the test suite matches the namespacing used by the production sources
(Tracking.* in Sources/RevenueCatAdMob/Tracking/).

File and class renames:
- RCAdMobTestCase.swift                       -> AdapterTestCase.swift
  (typealias RCAdMobTestCase -> AdapterTestCase)
- RCAdMobAPISurfaceTests.swift                -> APISurfaceTests.swift
- RCAdMobBannerDelegateForwardingTests.swift  -> BannerDelegateForwardingTests.swift
- RCAdMobBannerPaidHandlerBehaviorTests.swift -> BannerPaidHandlerBehaviorTests.swift
- RCAdMobBannerWrapperBehaviorTests.swift     -> BannerWrapperBehaviorTests.swift
- RCAdMobDelegateContractTests.swift          -> DelegateContractTests.swift
- RCAdMobFullScreenDelegateForwardingTests.swift
                                              -> FullScreenDelegateForwardingTests.swift
- RCAdMobNativeAdLoaderProxyBehaviorTests.swift
                                              -> NativeAdLoaderProxyBehaviorTests.swift
- RCAdMobNativeDelegateForwardingTests.swift  -> NativeDelegateForwardingTests.swift
- RCAdMobPrecisionMappingTests.swift          -> PrecisionMappingTests.swift
- RCAdMobRevenueConversionTests.swift         -> RevenueConversionTests.swift
- RCAdMobTrackingTests.swift                  -> AdapterTrackingTests.swift
  (also renames the secondary RCAdMobDelegateTrackingTests class within
   the file to AdapterDelegateTrackingTests)

Split RCAdMobDelegateRetentionTests.swift into per-store test files,
mirroring the store extractions:
- FullScreenDelegateStoreTests.swift — exercises Tracking.Adapter
  .fullScreenDelegateStore.retain(_:for:)
- NativeDelegateStoreTests.swift     — exercises Tracking.Adapter
  .nativeDelegateStore.retain(_:for:)

No production code changes. Verified with iOS Simulator test run
(87/87 passing).
