+++ AdapterSDKs/RevenueCatAdMob/Tests/RevenueCatAdMobTests/APISurfaceTests.swift
@@ -5,7 +5,7 @@ import GoogleMobileAds
 @_spi(Experimental) @testable import RevenueCatAdMob
 
 @available(iOS 15.0, *)
-final class RCAdMobAPISurfaceTests: RCAdMobTestCase {
+final class APISurfaceTests: AdapterTestCase {
 
     func testBannerAndNativeEntryPointsRemainAvailableInSwift() {
         let bannerLoadAndTrack: (GoogleMobileAds.BannerView) -> (
