+++ AdapterSDKs/RevenueCatAdMob/Tests/RevenueCatAdMobTests/AdapterTestCase.swift
@@ -1,4 +1,4 @@
 import XCTest
 
 /// Local base class for RevenueCatAdMob unit tests.
-typealias RCAdMobTestCase = XCTestCase
+typealias AdapterTestCase = XCTestCase
