+++ AdapterSDKs/RevenueCatAdMob/Tests/RevenueCatAdMobTests/NativeAdLoaderProxyBehaviorTests.swift
@@ -9,7 +9,7 @@ import GoogleMobileAds
 
 @available(iOS 15.0, *)
 // swiftlint:disable:next type_body_length
-final class RCAdMobNativeAdLoaderProxyBehaviorTests: RCAdMobTestCase {
+final class NativeAdLoaderProxyBehaviorTests: AdapterTestCase {
 
     func testLoadAndTrackInstallsProxyAndForwardsFailAndFinishLoading() {
         let adLoader = Self.makeAdLoader()
