+++ AdapterSDKs/RevenueCatAdMob/Tests/RevenueCatAdMobTests/PrecisionMappingTests.swift
@@ -6,7 +6,7 @@ import GoogleMobileAds
 @_spi(Experimental) @testable import RevenueCatAdMob
 
 @available(iOS 15.0, *)
-final class RCAdMobPrecisionMappingTests: RCAdMobTestCase {
+final class PrecisionMappingTests: AdapterTestCase {
 
     func testPreciseMapsToExact() {
         XCTAssertEqual(Tracking.Adapter.mapPrecision(.precise).rawValue, "exact")
